requirements.txt


dj-database-url==0.5.0
Django==2.1
django-allauth==0.41.0
django-crispy-forms==1.9.1
django-facebook-connect==1.0.1
django-heroku==0.3.1
django-storages==1.9.1
djangorestframework==3.10.1
docutils==0.15.2
fsspec==0.7.4
gunicorn==20.0.4
heroku==0.1.4
idna==2.8
jieba==0.42.1
jmespath==0.10.0
joblib==0.15.1
news-api==1.1.0
newsapi-python==0.2.6
numpy==1.18.4
oauthlib==3.1.0
pandas==0.25.0
psycopg2==2.8.5
PyMySQL==0.9.3
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0

scikit-learn==0.23.0
scipy==1.4.1
six==1.14.0
snownlp==0.12.3
soupsieve==2.0.1
threadpoolctl==2.0.0
true-encoding==1.0.2
urllib3==1.25.9